Polish Menubar and NewOfferPage
