+++ src/components/MenuBar.js
@@ -13,11 +13,14 @@ function MenuBar() {
 
   const [IsErrorDialogOpen, setIsErrorDialogOpen] = useState(false);
   const [errorMessage, setErrorMessage] = useState('');
+  const [currentUser, setCurrentUser] = useState(false);
   const [RedirectFunction, setRedirectFunction] = useState({});
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
   // const navigateToAdsPage = useCallback(() => navigate('/'), [navigate]);
-  const navigateBackToLogin = useCallback(() => navigate('/login'), [navigate]);
+  function navigateBackToLogin() {
+    navigate('/login');
+  }
   const hideCloseHandler = useCallback(() => setIsErrorDialogOpen(false), []);
 
   const logOut = useCallback(() => {
@@ -41,10 +44,17 @@ function MenuBar() {
       method: 'GET',
     }).then((reponse) => reponse.json().then((data) => {
       setIsLoggedIn(data.isuserloggedin);
-      if (!data.isuserloggedin && location.pathname !== '/login' && location.pathname !== '/signup') {
+      if (!data.isuserloggedin && location.pathname !== '/' && location.pathname !== '/login'
+        && location.pathname !== '/signup' && location.pathname !== '/ads' && location.pathname !== '/channels') {
         setErrorMessage('User is Logged out');
         setRedirectFunction(navigateBackToLogin);
         setIsErrorDialogOpen(true);
+      } else {
+        fetch('/get_current_user', {
+          method: 'GET',
+        }).then((userreponse) => userreponse.json().then((userdata) => {
+          setCurrentUser(userdata.current_user);
+        }));
       }
     }));
   });
@@ -54,7 +64,6 @@ function MenuBar() {
       <header className={classes.header}>
         <span className={classes.menu}>
           <Navbar.Brand href="/" style={{ color: 'black' }}>CraigsAdList</Navbar.Brand>
-          {isLoggedIn && (
           <Nav className="me-auto" activeKey="/ads">
             <Nav.Item>
               <Nav.Link href="/ads" style={{ color: 'black' }}>Find Ads</Nav.Link>
@@ -63,14 +72,13 @@ function MenuBar() {
               <Nav.Link href="/channels" style={{ color: 'black' }}>Find Channels</Nav.Link>
             </Nav.Item>
           </Nav>
-          )}
 
         </span>
         <div>
           <span className={classes.menu}>
-            <Button variant="outline-secondary" href="/new_add" style={{ fontWeight: 'bold' }}>+</Button>
+            {isLoggedIn && (<Button variant="outline-secondary" href="/new_add" style={{ fontWeight: 'bold' }}>+</Button>)}
 
-            <DropdownButton title="Menu" variant="secondary">
+            <DropdownButton title="Settings" variant="secondary">
               {!isLoggedIn && (
                 <div>
                   <Dropdown.Item>Not Logged In</Dropdown.Item>
@@ -81,8 +89,16 @@ function MenuBar() {
               )}
               {isLoggedIn && (
                 <div>
+
                   <MenuNavigation />
                   <Dropdown.Divider />
+
+                  <Dropdown.ItemText>
+                    Signed in as:
+                    {' '}
+                    <a href="/acount">{currentUser}</a>
+                  </Dropdown.ItemText>
+
                   <Dropdown.Item onClick={logOut}>Log out</Dropdown.Item>
                 </div>
               )}
