+++ src/pages/AdsPage.js
@@ -1,129 +1,12 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { useNavigate } from 'react-router-dom';
-import Card from 'react-bootstrap/Card';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Container from 'react-bootstrap/Container';
 import Form from 'react-bootstrap/Form';
 import InputGroup from 'react-bootstrap/InputGroup';
 import Button from 'react-bootstrap/Button';
-import Badge from 'react-bootstrap/Badge';
-
-function AdItem(props) {
-  const { ad } = props;
-  const {
-    id, creatorName, title, topics, text, reward,
-  } = ad;
-  const navigate = useNavigate();
-
-  function makeResponse() {
-    navigate('/new_response', { state: id });
-  }
-
-  return (
-    <Col>
-      <Card>
-        <Card.Body>
-          <Card.Title><h4><Badge pill bg="light" text="dark">{title}</Badge></h4></Card.Title>
-          <Card.Text>
-            <p>
-              Topics:
-              <br />
-              {topics}
-            </p>
-
-            <p>
-              Text:
-              <br />
-              {text}
-            </p>
-
-            <p>
-              <b style={{ float: 'left' }}>
-                Reward: $
-                {reward}
-              </b>
-              <b style={{ float: 'right' }}><Button variant="primary" onClick={() => makeResponse()}>Respond</Button></b>
-            </p>
-          </Card.Text>
-        </Card.Body>
-        <Card.Footer>
-          <small className="text-muted" style={{ float: 'left' }}>
-            #
-            {id}
-          </small>
-          <small className="text-muted" style={{ float: 'right' }}>
-            by
-            {' '}
-            {creatorName}
-          </small>
-        </Card.Footer>
-      </Card>
-    </Col>
-  );
-}
-AdItem.defaultProps = {
-  ad: PropTypes.shape({
-    id: 0,
-    creatorId: 0,
-    title: '',
-    topics: '',
-    text: '',
-    reward: 0,
-  }),
-};
-AdItem.propTypes = {
-  ad: PropTypes.shape({
-    id: PropTypes.number.isRequired,
-    creatorName: PropTypes.string.isRequired,
-    title: PropTypes.string.isRequired,
-    topics: PropTypes.string,
-    text: PropTypes.number.isRequired,
-    reward: PropTypes.number,
-  }),
-};
-
-function ListOfAds(props) {
-  const { query } = props;
-  const [ads, setAds] = useState(Array(0));
-
-  function getAds(newQuery) {
-    // fetch ads from database
-    fetch(`/return_ads?for=adsPage${newQuery}`, {
-      method: 'GET',
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        if (data.success) {
-          setAds(data.adsData);
-        } else {
-          throw new Error('Error while fetching ads data');
-        }
-      });
-  }
-
-  useEffect(() => { getAds(query); }, [props]);
-  if (ads.length === 0) {
-    return (
-      <h2 className="text-muted" align="center">
-        No ads are found. Try different filters?
-      </h2>
-    );
-  }
-  const listOfAds = ads.map((ad) => <AdItem ad={ad} />);
-  return (
-    <Row xs={1} md={2} className="g-4">
-      {listOfAds}
-    </Row>
-  );
-}
-ListOfAds.defaultProps = {
-  query: '',
-};
-ListOfAds.propTypes = {
-  query: PropTypes.string,
-};
+import ListOfAds from '../components/ListofAds';
 
 function SearchBar(props) {
   const { setQuery } = props;
