+++ src/pages/NewOfferPage.js
@@ -1,11 +1,20 @@
-import { useState, useEffect, useCallback } from 'react';
+import {
+  useState, useEffect, useCallback,
+} from 'react';
+import { useLocation } from 'react-router';
+import Card from 'react-bootstrap/Card';
+import Button from 'react-bootstrap/Button';
+import Col from 'react-bootstrap/Col';
+import Row from 'react-bootstrap/Row';
+import Form from 'react-bootstrap/Form';
+import InputGroup from 'react-bootstrap/InputGroup';
 import LoginErrorDialog from '../components/ui/js/LoginErrorDialog';
+import ListOfChannels from '../components/ListofChannels';
 
 function NewOfferPage() {
-  // const { state } = useLocation();
-  // const { selected_channel_id } = state;
+  const { state } = useLocation();
+  const { selectedId } = state;
 
-  const [ads, setAds] = useState([]);
   const [IsErrorDialogOpen, setIsErrorDialogOpen] = useState(false);
   const [ChannelName, setChannelName] = useState('');
   const [subscribers, setSubscibers] = useState('');
@@ -15,89 +24,59 @@ function NewOfferPage() {
   const showCloseHandler = useCallback(() => setIsErrorDialogOpen(true), []);
 
   useEffect(() => {
-    fetch('/return_channels?for=channelsPage', {
+    fetch(`/return_selected_channel?id=${selectedId}`, {
       method: 'GET',
     }).then((reponse) => reponse.json().then((data) => {
-      setChannelName(data.channels_data[0].channelName);
+      setChannelName(data.channelName);
       setSubscibers(data.channels_data[0].subscribers);
-      setPrice(data.channels_data[0].preferredReward);
-    }));
-  }, []);
-
-  useEffect(() => {
-    fetch('/return_ads', {
-      method: 'GET',
-    }).then((reponse) => reponse.json().then((data) => {
-      setAds(data.ads);
+      setPrice(data.preferredReward);
     }));
   }, []);
 
   return (
     <div>
-      <div style={{ padding: '5%', display: 'flex', justifyContent: 'space-between' }}>
-        <div>
-          <h>New Offer Page</h>
-          <div>Channel Info: </div>
+      <Row class>
+        <Col xs={8} md={4}>
           <div>
-            Channel Name:
-            {ChannelName}
-          </div>
-          <div>
-            Number of Subscribers:
-            {subscribers}
-          </div>
-          <div>Topic:</div>
-          <div>Prefered Contact:</div>
-          <div>
-            <div>
-              Prederred price of ads:
-              {`${price} `}
-              / 1k subscribers
+            <Card className="m-4" style={{ padding: '2rem' }}>
+              <Card.Title>New Offer Page</Card.Title>
+              <Card.Text>Channel Info: </Card.Text>
 
-            </div>
-          </div>
-          <div>
-            Message (optional):
-            <form>
-              <textarea
-                type="text"
-                style={{ resize: 'none', width: '30vw', height: '20vh' }}
-              />
+              <Card.Text>
+                Channel Name:
+                {ChannelName}
+              </Card.Text>
+              <Card.Text>
+                Number of Subscribers:
+                {subscribers}
+              </Card.Text>
+              <Card.Text>Topic:</Card.Text>
+              <Card.Text>Prefered Contact:</Card.Text>
 
-            </form>
-          </div>
-          <button type="button" onClick={showCloseHandler}>Make an Offer</button>
-        </div>
-        <div>
-          {ads.map((i) => (
-            <div>
+              <Card.Text>
+                Preferred price of ads:
+                <InputGroup className="mb-3">
+                  <InputGroup.Text id="basic-addon1">$</InputGroup.Text>
+                  <Form.Control name="maxReward" type="text" pattern="[0-9]*" value={price} />
+                </InputGroup>
+                / 1k subscribers
 
-              <div>
-                <input type="checkbox" />
-                creator_id:
-                {i.creator_id}
-              </div>
-              <div>
-                title:
-                {i.title}
-              </div>
-              <div>
-                topics:
-                {i.topics}
-              </div>
-              <div>
-                text:
-                {i.text}
-              </div>
-              <div>
-                reward:
-                {i.reward}
-              </div>
-              <br />
-            </div>
-          ))}
-        </div>
-      </div>
+              </Card.Text>
+
+              <Card.Text>Message (optional):</Card.Text>
+              <Form>
+                <Form.Control as="textarea" rows={3} />
+
+              </Form>
+
+              <Button variant="outline-secondary" onClick={showCloseHandler}>Make an Offer</Button>
+            </Card>
+          </div>
+        </Col>
+        <Col className="m-4">
+          <ListOfChannels />
+        </Col>
+      </Row>
       {IsErrorDialogOpen && (
       <LoginErrorDialog
         message="Placeholder. Will complete when Email processing is implemented"
